kynbiot-v2.3/coordinator/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
psycopg[binary]==3.2.3
redis==5.2.1
pydantic-settings==2.7.1
python-jose==3.3.0
